requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.0.76
easyocr==1.6.2
Pillow==10.0.0
numpy==1.24.3
